+++ dev/js/Network.js
@@ -253,7 +253,7 @@ class Network {
         }
     }
 
-    train (dataSet, {epochs=1, callback, log=true, miniBatchSize=1, shuffle=false, validation}={}) {
+    train (dataSet, {epochs=1, callback, callbackInterval=1, log=true, miniBatchSize=1, shuffle=false, validation}={}) {
 
         this.miniBatchSize = typeof miniBatchSize=="boolean" && miniBatchSize ? dataSet[0].expected.length : miniBatchSize
         this.validation = validation
@@ -381,7 +381,7 @@ class Network {
 
                 elapsed = Date.now() - startTime
 
-                if (typeof callback=="function") {
+                if ((iterationIndex%callbackInterval == 0 || validationError) && typeof callback=="function") {
                     callback({
                         iterations: this.iterations,
                         validations: this.validations,
@@ -391,7 +391,12 @@ class Network {
                 }
 
                 if (iterationIndex < dataSet.length) {
-                    setTimeout(doIteration.bind(this), 0)
+
+                    if (iterationIndex%callbackInterval == 0) {
+                        setTimeout(doIteration.bind(this), 0)
+                    } else {
+                        doIteration()
+                    }
 
                 } else {
                     epochsCounter++